+++ weblate/machinery/modernmt.py
@@ -22,11 +22,9 @@ import json
 from django.conf import settings
 
 import weblate.utils.version
-from weblate.machinery.base import (
-    MachineTranslation,
-    MachineTranslationError,
-    MissingConfiguration,
-)
+
+from .base import MachineTranslation, MachineTranslationError
+from .forms import ModernMTMachineryForm
 
 
 class ModernMTTranslation(MachineTranslation):
@@ -34,17 +32,19 @@ class ModernMTTranslation(MachineTranslation):
 
     name = "ModernMT"
     max_score = 90
+    settings_form = ModernMTMachineryForm
 
-    def __init__(self):
-        """Check configuration."""
-        super().__init__()
-        if settings.MT_MODERNMT_KEY is None:
-            raise MissingConfiguration("ModernMT requires API key")
+    @staticmethod
+    def migrate_settings():
+        return {
+            "key": settings.MT_MODERNMT_KEY,
+            "url": settings.MT_MODERNMT_URL,
+        }
 
     def get_authentication(self):
         """Hook for backends to allow add authentication headers to request."""
         return {
-            "MMT-ApiKey": settings.MT_MODERNMT_KEY,
+            "MMT-ApiKey": self.settings["key"],
             "MMT-Platform": "Weblate",
             "MMT-PlatformVersion": weblate.utils.version.VERSION,
         }
@@ -55,7 +55,7 @@ class ModernMTTranslation(MachineTranslation):
 
     def download_languages(self):
         """List of supported languages."""
-        response = self.request("get", settings.MT_MODERNMT_URL + "languages")
+        response = self.request("get", self.settings["url"] + "languages")
         payload = response.json()
 
         for source, targets in payload["data"].items():
@@ -74,7 +74,7 @@ class ModernMTTranslation(MachineTranslation):
         """Download list of possible translations from a service."""
         response = self.request(
             "get",
-            settings.MT_MODERNMT_URL + "translate",
+            self.settings["url"] + "translate",
             params={"q": text, "source": source, "target": language},
         )
         payload = response.json()