+++ weblate/machinery/tests.py
@@ -27,14 +27,16 @@ from urllib.parse import parse_qs
 import responses
 from botocore.stub import ANY, Stubber
 from django.test import TestCase
-from django.test.utils import override_settings
+from django.urls import reverse
 from google.cloud.translate_v3 import (
     SupportedLanguages,
     TranslateTextResponse,
     TranslationServiceClient,
 )
 
+import weblate.machinery.models
 from weblate.checks.tests.test_checks import MockUnit
+from weblate.configuration.models import Setting
 from weblate.machinery.apertium import ApertiumAPYTranslation
 from weblate.machinery.aws import AWSTranslation
 from weblate.machinery.baidu import BAIDU_API, BaiduTranslation
@@ -42,7 +44,6 @@ from weblate.machinery.base import (
     MachineryRateLimit,
     MachineTranslation,
     MachineTranslationError,
-    MissingConfiguration,
 )
 from weblate.machinery.deepl import DeepLTranslation
 from weblate.machinery.dummy import DummyTranslation
@@ -63,9 +64,10 @@ from weblate.machinery.tmserver import AMAGAMA_LIVE, AmagamaTranslation
 from weblate.machinery.weblatetm import WeblateTranslation
 from weblate.machinery.yandex import YandexTranslation
 from weblate.machinery.youdao import YoudaoTranslation
-from weblate.trans.models.unit import Unit
+from weblate.trans.models import Project, Unit
 from weblate.trans.tests.test_views import FixtureTestCase
 from weblate.trans.tests.utils import get_test_file
+from weblate.utils.classloader import load_class
 from weblate.utils.db import using_postgresql
 from weblate.utils.state import STATE_TRANSLATED
 
@@ -263,7 +265,8 @@ TERMINOLOGY_TRANSLATE = b"""
 """
 TERMINOLOGY_WDSL = get_test_file("microsoftterminology.wsdl")
 
-GOOGLEV3_KEY = get_test_file("googlev3.json")
+with open(get_test_file("googlev3.json")) as handle:
+    GOOGLEV3_KEY = handle.read()
 
 DEEPL_RESPONSE = {"translations": [{"detected_source_language": "EN", "text": "Hallo"}]}
 DEEPL_LANG_RESPONSE = [
@@ -308,9 +311,10 @@ class BaseMachineTranslationTest(TestCase):
     SOURCE_BLANK = "Hello"
     SOURCE_TRANSLATED = "Hello, world!"
     EXPECTED_LEN = 2
+    CONFIGURATION = {}
 
     def get_machine(self, cache=False):
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         machine.cache_translations = cache
         return machine
@@ -456,7 +460,7 @@ class GlosbeTranslationTest(BaseMachineTranslationTest):
         # This raises an exception
         self.test_error()
         # The second call should not perform due to rate limiting being cached
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         self.assert_translate(
             self.SUPPORTED, self.SOURCE_TRANSLATED, 0, machine=machine
         )
@@ -464,7 +468,7 @@ class GlosbeTranslationTest(BaseMachineTranslationTest):
     @responses.activate
     def test_ratelimit_set(self):
         """Test manual setting of rate limit."""
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         machine.set_rate_limit()
         self.assert_translate(
@@ -472,11 +476,15 @@ class GlosbeTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_MYMEMORY_EMAIL="test@weblate.org")
 class MyMemoryTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = MyMemoryTranslation
     EXPECTED_LEN = 3
     NOTSUPPORTED = "ia"
+    CONFIGURATION = {
+        "email": "test@weblate.org",
+        "username": "user",
+        "key": "key",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -490,12 +498,14 @@ class MyMemoryTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_APERTIUM_APY="http://apertium.example.com/")
 class ApertiumAPYTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = ApertiumAPYTranslation
     ENGLISH = "eng"
     SUPPORTED = "spa"
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "url": "http://apertium.example.com",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -525,7 +535,7 @@ class ApertiumAPYTranslationTest(BaseMachineTranslationTest):
     @responses.activate
     def test_translations_cache(self):
         self.mock_response()
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         self.assert_translate(
             self.SUPPORTED, self.SOURCE_TRANSLATED, self.EXPECTED_LEN, machine=machine
@@ -533,17 +543,22 @@ class ApertiumAPYTranslationTest(BaseMachineTranslationTest):
         self.assertEqual(len(responses.calls), 2)
         responses.reset()
         # New instance should use cached languages and translations
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         self.assert_translate(
             self.SUPPORTED, self.SOURCE_TRANSLATED, self.EXPECTED_LEN, machine=machine
         )
         self.assertEqual(len(responses.calls), 0)
 
 
-@override_settings(MT_MICROSOFT_COGNITIVE_KEY="KEY")
 class MicrosoftCognitiveTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = MicrosoftCognitiveTranslation
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "key": "KEY",
+        "endpoint_url": "api.cognitive.microsoft.com",
+        "base_url": "api.cognitive.microsofttranslator.com",
+        "region": "",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -572,8 +587,14 @@ class MicrosoftCognitiveTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_MICROSOFT_COGNITIVE_KEY="KEY", MT_MICROSOFT_REGION="westeurope")
 class MicrosoftCognitiveTranslationRegionTest(MicrosoftCognitiveTranslationTest):
+    CONFIGURATION = {
+        "key": "KEY",
+        "endpoint_url": "api.cognitive.microsoft.com",
+        "base_url": "api.cognitive.microsofttranslator.com",
+        "region": "westeurope",
+    }
+
     def mock_response(self):
         responses.add(
             responses.POST,
@@ -645,10 +666,12 @@ class MicrosoftTerminologyServiceTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_GOOGLE_KEY="KEY")
 class GoogleTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = GoogleTranslation
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "key": "KEY",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -680,7 +703,7 @@ class GoogleTranslationTest(BaseMachineTranslationTest):
     @responses.activate
     def test_ratelimit_set(self):
         """Test manual setting of rate limit."""
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         machine.set_rate_limit()
         self.assert_translate(
@@ -688,17 +711,14 @@ class GoogleTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(
-    MT_GOOGLE_CREDENTIALS=GOOGLEV3_KEY, MT_GOOGLE_PROJECT="translating-7586"
-)
 class GoogleV3TranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = GoogleV3Translation
     EXPECTED_LEN = 1
-
-    @override_settings(MT_GOOGLE_CREDENTIALS=None)
-    def test_google_apiv3_bad_config(self):
-        with self.assertRaises(MissingConfiguration):
-            self.get_machine()
+    CONFIGURATION = {
+        "project": "translating-7586",
+        "location": "global",
+        "credentials": GOOGLEV3_KEY,
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -766,10 +786,12 @@ class AmagamaTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_YANDEX_KEY="KEY")
 class YandexTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = YandexTranslation
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "key": "KEY",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -815,13 +837,16 @@ class YandexTranslationTest(BaseMachineTranslationTest):
             self.assert_translate(self.SUPPORTED, self.SOURCE_BLANK, 0)
 
 
-@override_settings(MT_YOUDAO_ID="id", MT_YOUDAO_SECRET="secret")
 class YoudaoTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = YoudaoTranslation
     EXPECTED_LEN = 1
     SUPPORTED = "de"
     NOTSUPPORTED = "cs"
     ENGLISH = "EN"
+    CONFIGURATION = {
+        "key": "id",
+        "secret": "secret",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -839,11 +864,11 @@ class YoudaoTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_NETEASE_KEY="key", MT_NETEASE_SECRET="secret")
 class NeteaseSightTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = NeteaseSightTranslation
     EXPECTED_LEN = 1
     SUPPORTED = "zh"
+    CONFIGURATION = {"key": "id", "secret": "secret"}
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -862,11 +887,14 @@ class NeteaseSightTranslationTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_BAIDU_ID="id", MT_BAIDU_SECRET="secret")
 class BaiduTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = BaiduTranslation
     EXPECTED_LEN = 1
     NOTSUPPORTED = "ia"
+    CONFIGURATION = {
+        "key": "id",
+        "secret": "secret",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -900,10 +928,16 @@ class BaiduTranslationTest(BaseMachineTranslationTest):
             self.assert_translate(self.SUPPORTED, self.SOURCE_TRANSLATED, 0)
 
 
-@override_settings(MT_SAP_BASE_URL="http://sth.example.com/")
 class SAPTranslationHubTest(BaseMachineTranslationTest):
     MACHINE_CLS = SAPTranslationHub
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "key": "",
+        "username": "",
+        "password": "",
+        "enable_mt": False,
+        "url": "http://sth.example.com/",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -937,10 +971,23 @@ class SAPTranslationHubTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(MT_MODERNMT_KEY="key")
+class SAPTranslationHubAuthTest(SAPTranslationHubTest):
+    CONFIGURATION = {
+        "key": "id",
+        "username": "username",
+        "password": "password",
+        "url": "http://sth.example.com/",
+        "enable_mt": False,
+    }
+
+
 class ModernMTHubTest(BaseMachineTranslationTest):
     MACHINE_CLS = ModernMTTranslation
     EXPECTED_LEN = 1
+    CONFIGURATION = {
+        "key": "KEY",
+        "url": "https://api.modernmt.com/",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -986,22 +1033,16 @@ class ModernMTHubTest(BaseMachineTranslationTest):
         )
 
 
-@override_settings(
-    MT_SAP_SANDBOX_APIKEY="http://sandbox.example.com",
-    MT_SAP_USERNAME="username",
-    MT_SAP_PASSWORD="password",
-)
-class SAPTranslationHubAuthTest(SAPTranslationHubTest):
-    pass
-
-
-@override_settings(MT_DEEPL_KEY="KEY")
 class DeepLTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = DeepLTranslation
     EXPECTED_LEN = 1
     ENGLISH = "EN"
     SUPPORTED = "DE"
     NOTSUPPORTED = "CS"
+    CONFIGURATION = {
+        "key": "KEY",
+        "url": "https://api.deepl.com/v2/",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -1043,7 +1084,7 @@ class DeepLTranslationTest(BaseMachineTranslationTest):
             self.assertIn("formality", payload)
             return (200, headers, json.dumps(DEEPL_RESPONSE))
 
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         self.mock_languages()
         responses.add_callback(
@@ -1067,7 +1108,7 @@ class DeepLTranslationTest(BaseMachineTranslationTest):
             self.assertEqual(payload["text"], ['Hello, <x id="7"></x>!'])
             return (200, headers, json.dumps(DEEPL_RESPONSE))
 
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         self.mock_languages()
         responses.add_callback(
@@ -1086,7 +1127,7 @@ class DeepLTranslationTest(BaseMachineTranslationTest):
 
     @responses.activate
     def test_cache(self):
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         self.mock_response()
         # Fetch from service
@@ -1096,20 +1137,23 @@ class DeepLTranslationTest(BaseMachineTranslationTest):
         self.assertEqual(len(responses.calls), 2)
         responses.reset()
         # Fetch from cache
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         self.assert_translate(
             self.SUPPORTED, self.SOURCE_TRANSLATED, self.EXPECTED_LEN, machine=machine
         )
         self.assertEqual(len(responses.calls), 0)
 
 
-@override_settings(MT_LIBRETRANSLATE_API_URL="https://libretranslate.com/")
 class LibreTranslateTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = LibreTranslateTranslation
     EXPECTED_LEN = 1
     ENGLISH = "en"
     SUPPORTED = "es"
     NOTSUPPORTED = "cs"
+    CONFIGURATION = {
+        "url": "https://libretranslate.com/",
+        "key": "",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -1136,7 +1180,7 @@ class LibreTranslateTranslationTest(BaseMachineTranslationTest):
 
     @responses.activate
     def test_cache(self):
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         machine.delete_cache()
         self.mock_response()
         # Fetch from service
@@ -1146,18 +1190,22 @@ class LibreTranslateTranslationTest(BaseMachineTranslationTest):
         self.assertEqual(len(responses.calls), 2)
         responses.reset()
         # Fetch from cache
-        machine = self.MACHINE_CLS()
+        machine = self.MACHINE_CLS(self.CONFIGURATION)
         self.assert_translate(
             self.SUPPORTED, self.SOURCE_TRANSLATED, self.EXPECTED_LEN, machine=machine
         )
         self.assertEqual(len(responses.calls), 0)
 
 
-@override_settings(MT_AWS_REGION="us-west-2")
 class AWSTranslationTest(BaseMachineTranslationTest):
     MACHINE_CLS = AWSTranslation
     EXPECTED_LEN = 1
     NOTSUPPORTED = "ia"
+    CONFIGURATION = {
+        "region": "us-west-2",
+        "key": "key",
+        "secret": "secret",
+    }
 
     def mock_empty(self):
         raise SkipTest("Not tested")
@@ -1236,7 +1284,7 @@ class WeblateTranslationTest(FixtureTestCase):
         return super()._databases_support_transactions()
 
     def test_empty(self):
-        machine = WeblateTranslation()
+        machine = WeblateTranslation({})
         results = machine.translate(self.get_unit(), self.user)
         self.assertEqual(results, [])
 
@@ -1249,6 +1297,129 @@ class WeblateTranslationTest(FixtureTestCase):
         other.state = STATE_TRANSLATED
         other.save()
         # Perform lookup
-        machine = WeblateTranslation()
+        machine = WeblateTranslation({})
         results = machine.translate(unit, self.user)
         self.assertNotEqual(results, [])
+
+
+class ViewsTest(FixtureTestCase):
+    """Testing of AJAX/JS views."""
+
+    @staticmethod
+    def ensure_dummy_mt():
+        """Ensure we have dummy mt installed."""
+        name = "weblate.machinery.dummy.DummyTranslation"
+        service = load_class(name, "TEST")
+        if service.get_identifier() not in weblate.machinery.models.MACHINERY:
+            weblate.machinery.models.MACHINERY[service.get_identifier()] = service
+        Setting.objects.create(
+            category=Setting.CATEGORY_MT, name=service.get_identifier(), value={}
+        )
+        return service
+
+    def test_translate(self):
+        self.ensure_dummy_mt()
+        unit = self.get_unit()
+        response = self.client.post(
+            reverse("js-translate", kwargs={"unit_id": unit.id, "service": "dummy"})
+        )
+        self.assertContains(response, "Ahoj")
+        data = response.json()
+        self.assertEqual(
+            data["translations"],
+            [
+                {
+                    "quality": 100,
+                    "service": "Dummy",
+                    "text": "Nazdar světe!",
+                    "source": "Hello, world!\n",
+                },
+                {
+                    "quality": 100,
+                    "service": "Dummy",
+                    "text": "Ahoj světe!",
+                    "source": "Hello, world!\n",
+                },
+            ],
+        )
+
+        # Invalid service
+        response = self.client.post(
+            reverse("js-translate", kwargs={"unit_id": unit.id, "service": "invalid"})
+        )
+        self.assertEqual(response.status_code, 404)
+
+    def test_memory(self):
+        unit = self.get_unit()
+        url = reverse("js-memory", kwargs={"unit_id": unit.id})
+        # Missing param
+        response = self.client.post(url)
+        self.assertEqual(response.status_code, 400)
+        # Valid query
+        response = self.client.post(url, {"q": "a"})
+        self.assertEqual(response.status_code, 200)
+        data = response.json()
+        self.assertEqual(data["service"], "Weblate Translation Memory")
+
+    def test_configure_global(self):
+        service = self.ensure_dummy_mt()
+        list_url = reverse("machinery-list")
+        edit_url = reverse(
+            "machinery-edit", kwargs={"machinery": service.get_identifier()}
+        )
+        response = self.client.get(list_url)
+        self.assertEqual(response.status_code, 403)
+
+        self.user.is_superuser = True
+        self.user.save()
+
+        response = self.client.get(list_url)
+        self.assertContains(response, edit_url)
+
+        self.client.post(edit_url, {"delete": "1"})
+        self.assertFalse(
+            Setting.objects.filter(
+                category=Setting.CATEGORY_MT, name=service.get_identifier()
+            ).exists()
+        )
+        self.client.post(edit_url, {"install": "1"})
+        self.assertTrue(
+            Setting.objects.filter(
+                category=Setting.CATEGORY_MT, name=service.get_identifier()
+            ).exists()
+        )
+
+    def test_configure_project(self):
+        service = self.ensure_dummy_mt()
+        list_url = reverse("machinery-list", kwargs={"project": self.project.slug})
+        edit_url = reverse(
+            "machinery-edit",
+            kwargs={
+                "project": self.project.slug,
+                "machinery": service.get_identifier(),
+            },
+        )
+        response = self.client.get(list_url)
+        self.assertEqual(response.status_code, 403)
+
+        self.make_manager()
+
+        response = self.client.get(list_url)
+        self.assertContains(response, edit_url)
+
+        self.client.post(edit_url, {"delete": "1"})
+        self.assertTrue(
+            Setting.objects.filter(
+                category=Setting.CATEGORY_MT, name=service.get_identifier()
+            ).exists()
+        )
+        project = Project.objects.get(pk=self.project.id)
+        self.assertEqual(project.machinery_settings["dummy"], None)
+        self.client.post(edit_url, {"enable": "1"})
+        self.assertTrue(
+            Setting.objects.filter(
+                category=Setting.CATEGORY_MT, name=service.get_identifier()
+            ).exists()
+        )
+        project = Project.objects.get(pk=self.component.project_id)
+        self.assertNotIn("dummy", project.machinery_settings)